Tests/LibWeb/Text/expected/XHR/XMLHttpRequest-response-header-decoding.txt

getAllResponseHeaders()
refresh: 0;./refreshed.txt?ÿ

getResponseHeader("Refresh") => '0;./refreshed.txt?ÿ'
